runtime: split out the SPPF into a `forest` module.
