+++ src/generate/templates/imports.rs
@@ -1,5 +1,5 @@
+use gll::forest::{nd::Arrow, traverse, ParseNode, ParseNodeKind};
 use gll::parse_node::ParseNodeShape;
-use gll::runtime::{nd::Arrow, traverse, CodeLabel, ParseNode, ParseNodeKind};
 use std::any;
 use std::fmt;
 use std::marker::PhantomData;
